TopologicalSortComparator: Changed to exclude any abstraction definition from its graph that is not in the collection of abstraction definitions passed into the constructor. This should fix an issue where the graph was put into an inconsistent state with abstraction definitions that had inverseIsA and abstractionInto relationships at the same time.
JBossRuleCreator: Only add a combiner rule to high-level and low-level abstraction definitions when an abstractedFrom relationship is present. This avoids an issue with rule salience that emerged from the above fix to the sorter that could cause abstractions to get computed incorrectly. For high-level and low-level abstraction definitions with no abstractedFrom relationship, it makes no sense semantically to attempt to attempt to combine them.

AbstractionDefinition: Removed the spurious "public abstract" from the interface's methods.